Add Chiliz fan token staking TVL calculation (#13570)
