+++ projects/chiliz-fan-token-staking/index.js
@@ -1,0 +1,5 @@
+const { getExport } = require('../chiliz-peppercoin/util');
+const STAKING_CONTRACT = '0x5ff7f9724fd477d9a07dcdb894d0ca7f8fae1501';
+
+
+module.exports = getExport(STAKING_CONTRACT, 20248272)
